Progress
Added a application settings module that is added to OBDForms and Projects, this centralizes the ability to read/write registry values and automatically parses command line parameters - which can be usefull for debug modes etc. Also updated the OBD Form to improve the repainting by creating lists with our OBD components on form creation, which means we dont need to loop over all components on every window state change.