Progress
Added about box and splash screen info. Created a OBD Datamodule that has access to the same properties as the OBD Form, and created a wizard for the OBD Datamodule and expanded the project wizard to include a OBD Datamodule.

Also started on creating a VIN parser (needs to be tested) which parses basic info like Region, Country, Manufacturer and (possible) year(s). This can be extended later but should provide enough info to be able to decode the VIN retrieved from Mode $09. I will create additional components for getting more info out of a VIN through a (paid) API.